# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin panel API",
        "description": "Codeyad-React Learning-Ecommerce admin panel API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get users",
                "description": "get all users or with pagination",
                "operationId": "getUsers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "searchChar",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "[{}, {}]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Add users",
                "description": "store one users",
                "operationId": "addUser",
                "requestBody": {
                    "description": "add one category",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_name",
                                    "phone",
                                    "password"
                                ],
                                "properties": {
                                    "user_name": {
                                        "type": "string",
                                        "example": "test username"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "test name"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "test family"
                                    },
                                    "phone": {
                                        "type": "number",
                                        "example": "09110000011"
                                    },
                                    "national_code": {
                                        "type": "number",
                                        "example": "1111111111"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "testemail@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "example": "2000-10-10"
                                    },
                                    "gender": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "roles_id": {
                                        "type": "object",
                                        "example": "[2,3]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "[{}, {}]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/admin/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get one users",
                "description": "Get one user with id",
                "operationId": "oneUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "{}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Edit user",
                "description": "Edit one user",
                "operationId": "editUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "edit category",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_name": {
                                        "type": "string",
                                        "example": "test username edited"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "test name edited"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "test family edited"
                                    },
                                    "phone": {
                                        "type": "number",
                                        "example": "09110000012"
                                    },
                                    "national_code": {
                                        "type": "number",
                                        "example": "2222222222"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "testemail_edited@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "example": "2000-11-10"
                                    },
                                    "gender": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "roles_id": {
                                        "type": "object",
                                        "example": "[4,5]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "عملیت با موفقیت انجام شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user",
                "description": "Delete one user",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "حذف با موفقیت انجام شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}